fixed repeating health plan alarms, show activities for 3 days, show comments for care giver, allow decimals in measurement values

+++ netcare-core/src/main/java/org/callistasoftware/netcare/core/api/Measurement.java
@@ -33,21 +33,21 @@ public interface Measurement {
 	/**
 	 * Returns target if not an interval.
 	 */
-	int getTarget();
+	float getTarget();
 
 	/**
 	 * Returns min target if there is an interval.
 	 */
-	int getMinTarget();
+	float getMinTarget();
 
 	/**
 	 * Returns max target if there is an interval.
 	 * 
 	 */
-	int getMaxTarget();
+	float getMaxTarget();
 	
 	/**
 	 * Returns reported value.
 	 */
-	int getReportedValue();
+	float getReportedValue();
 }
